feat: create module server cert and key via edgeCA (#20)
In the course of implementing the automatic generation of server certificates using edgeCA, we had to discontinue support for injecting the app via omnect-cli.
As a result, all artifacts related to systemd handling were removed from the repo.

For local development use the build-and-run.sh file and read the precondition based on this script.

Signed-off-by: Joerg Zeidler <62105035+JoergZeidler@users.noreply.github.com>